Move Reader into its own internal module.
