+++ fused-effects.cabal
@@ -91,6 +91,7 @@ library
     Control.Effect.Internal
     Control.Effect.Lift.Internal
     Control.Effect.NonDet.Internal
+    Control.Effect.Reader.Internal
   build-depends:
       base           >= 4.9 && < 4.14
     , deepseq       ^>= 1.4.3
