+++ src/Control/Effect/Internal.hs
@@ -2,7 +2,6 @@
 module Control.Effect.Internal
 ( -- * Effects
   Error
-, Reader(..)
 , State(..)
 , Throw(..)
 , Writer(..)
@@ -15,22 +14,6 @@ import GHC.Generics (Generic1)
 
 -- | @since 0.1.0.0
 type Error e = Throw e :+: Catch e
-
-
--- | @since 0.1.0.0
-data Reader r m k
-  = Ask (r -> m k)
-  | forall b . Local (r -> r) (m b) (b -> m k)
-
-deriving instance Functor m => Functor (Reader r m)
-
-instance HFunctor (Reader r) where
-  hmap f (Ask k)       = Ask           (f . k)
-  hmap f (Local g m k) = Local g (f m) (f . k)
-
-instance Effect (Reader r) where
-  handle state handler (Ask k)       = Ask (handler . (<$ state) . k)
-  handle state handler (Local f m k) = Local f (handler (m <$ state)) (handler . fmap k)
 
 
 -- | @since 0.1.0.0
