+++ src/Control/Effect/Reader.hs
@@ -26,7 +26,7 @@ module Control.Effect.Reader
 ) where
 
 import Control.Algebra
-import Control.Effect.Internal (Reader(..))
+import Control.Effect.Reader.Internal (Reader(..))
 
 -- | Retrieve the environment value.
 --
